Handle only int, not float as described in the ddb

+++ Back/src/main/java/com/openclassrooms/RentalProject/DTO/RentalDto.java
@@ -8,8 +8,8 @@ import lombok.Data;
 public class RentalDto {
 	private int id;
 	private String name;
-	private float surface;
-	private float price;
+	private int surface;
+	private int price;
 	private String picture;
 	private String description;
 	private int owner_id;
